Split the memory module client to stay under the layout gate
The loading-state work pushed two files past the 750-line limit
`check-openhuman-rust-layout.mjs` enforces: `memory_part_01.rs` to 842 and
`memory_tests.rs` to 767.

Both follow the split conventions already in the tree. `memory_part_04.rs`
takes the `MemoryCore` / `MemoryRecall` / `MemoryPortability` / `MemoryIngest`
impls and is `include!`d after the other three parts, so it is concatenated
into the same module and needs no imports of its own. The `module_call!` and
`module_call_slow!` macros stay in `memory_part_01.rs`: a `macro_rules!` is
only in scope for text that follows it, and `memory_part_02.rs` and
`memory_part_03.rs` are concatenated before part four.

`memory_tests_part_01_tests.rs` takes the real-module round-trip test and the
scoring accessor test, wired with `#[path] mod part_01_tests;` the way
`ops_tests_part_NN_tests.rs` already is in the composio domain. That file is a
submodule rather than an `include!`, so it opens with `use super::*;` and the
one `super::capabilities_for` call becomes `super::super::capabilities_for` —
it sits a level deeper than it did.

No behaviour change: same 116 tests pass, and the layout gate is clean.
